+++ CHANGES.md
@@ -49,6 +49,7 @@ To be released.
  -  (Libplanet.Net) Added
     `InvalidProposerProposeMessageException` class.  [[#PBFT]]
  -  (Libplanet.Net) Added `Message.GetHash()` method.  [[#PBFT]]
+ -  (Libplanet.Net) Added `Gossip` class.  [[#PBFT]]
 
 ### Behavioral changes